+++ s3/src/test/scala/S3Spec.scala
@@ -33,23 +33,6 @@ class S3Spec extends FlatSpec with Matchers with ScalaFutures with BeforeAndAfte
     new com.amazonaws.auth.profile.ProfileCredentialsProvider("mfg")
   )().materialized(fm)
 
-  it should "upload/list/delete small files" in {
-    s3Client.deleteObjects(bucket, s"$keyPrefix").futureValue
-    s3Client.putObject(bucket, s"$keyPrefix/small.txt", new java.io.File(getClass.getResource("/small.txt").getPath)).futureValue
-    val l = s3Client.listFiles(bucket, Some(keyPrefix)).futureValue
-    s3Client.deleteObjects(bucket, s"$keyPrefix/small.txt").futureValue
-    val l2 = s3Client.listFiles(bucket, Some(keyPrefix)).futureValue
-
-    (l map (_.getKey)) should equal(List(s"$keyPrefix/small.txt"))
-    l2 should be('empty)
-  }
-
-  it should "throw an exception when a file is non-existent" in {
-    val source = s3Client.getFileAsStream(bucket, "foo")
-    intercept[AmazonS3Exception] {
-      Await.result(source.runFold(ByteString.empty)(_ ++ _).map(_.compact), 5 seconds)
-    }
-  }
 
   it should "upload and download a big file as a single file" in {
     val futBytes = StreamConverters.fromInputStream(() => getClass.getResourceAsStream("/big.txt"))
@@ -58,6 +41,8 @@ class S3Spec extends FlatSpec with Matchers with ScalaFutures with BeforeAndAfte
       .runFold(ByteString.empty)(_ ++ _)
       .map(_.compact)
 
+    whenReady(futBytes) { case _ => () }
+
     val expectedBytes = StreamConverters.fromInputStream(() => getClass.getResourceAsStream("/big.txt"))
       .runFold(ByteString.empty)(_ ++ _).map(_.compact)
 
@@ -65,42 +50,6 @@ class S3Spec extends FlatSpec with Matchers with ScalaFutures with BeforeAndAfte
       bytes shouldEqual expectedBytes
     }
   }
-
-  it should "download a big file and chunk it by line" in {
-    val futLines = s3Client.getFileAsStream(bucket, s"$keyPrefix/big")
-      .via(FlowExt.rechunkByteStringBySize(2 * 1024 * 1024))
-      .via(FlowExt.rechunkByteStringBySeparator(ByteString("\n"), 8 * 1024))
-      .map(_.utf8String)
-      .runWith(Sink.seq)
-
-    val futExpectedLines = StreamConverters
-        .fromInputStream(() => getClass.getResourceAsStream("/big.txt"))
-        .runFold(ByteString.empty)(_ ++ _)
-        .map(_.compact.utf8String)
-        .map(_.split("\n").to[scala.collection.immutable.Seq])
-
-    whenReady(futLines zip futExpectedLines) { case (lines, expectedLines) =>
-      lines shouldEqual expectedLines
-    }
-  }
-
-
-  it should "upload and download a big file as a multipart file" in {
-    val bytes = StreamConverters
-      .fromInputStream(() => getClass.getResourceAsStream("/big.txt"), chunkSize = 2 * 1024 * 1024)
-      .via(s3Client.uploadStreamAsMultipartFile(bucket, s"$keyPrefix/big", nbChunkPerFile = 1, chunkUploadConcurrency = 2))
-      .via(FlowExt.fold[CompleteMultipartUploadResult, Vector[CompleteMultipartUploadResult]](Vector.empty)(_ :+ _))
-      .flatMapConcat(_ => s3Client.getMultipartFileAsStream(bucket, s"$keyPrefix/big.part"))
-      .runFold(ByteString.empty)(_ ++ _)
-      .map(_.compact).futureValue
-
-   val expectedBytes = StreamConverters
-      .fromInputStream(() => getClass.getResourceAsStream("/big.txt"))
-      .runFold(ByteString.empty)(_ ++ _).map(_.compact).futureValue
-
-    bytes shouldEqual expectedBytes
-  }
-
 
   override def afterAll() = {
     s3Client.shutdown()